backend/requirements.txt
fastapi
uvicorn[standard]
pdfminer.six
python-docx
openai
python-dotenv
python-multipart
supabase
pydantic_settings
